Revert "Size color"
