+++ client/src/App.js
@@ -1,0 +1,7 @@
+function App() {
+  return (
+    <h1 className="text-3xl font-bold underline">Hello world!</h1>
+  );
+}
+
+export default App;
