--- products.json
@@ -1,26 +1,0 @@
-{
-    "products": [
-      {
-        "id": 1,
-        "title": "Hoodies",
-        "description": "Monk Cotton Hoodies",
-        "more": "Unisex and can be worn during any weather",
-        "price": 2200,
-        "image": "http://localhost:3000/images/purchasehoodie.png",
-        "size" : ["Small", "Medium", "Large", "X-Large"],
-        "color" : ["Blue", "Yellow", "White", "Black"]
-      },
-
-      {
-        "id": 2,
-        "title": "Shorts",
-        "description": "Monk Cotton Shorts",
-        "more": "Comfortable and ideal for sunny",
-        "price": 549,
-        "image": "http://localhost:3000/images/short.png",
-        "size" : ["Small", "Medium", "Large", "X-Large"],
-        "color" : ["Blue", "White", "Black"]
-      }
-      
-    ]
-  }
