--- src/pages/AboutUs.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const AboutUs = () => {
-  return (
-    <div>AboutUs</div>
-  )
-}
-
-export default AboutUs
